ref: rename algorithms to models, and models to instances in user facing parts
