--- src/app/@breadcrumbs/algorithms/[slug]/page.tsx
@@ -1,9 +1,0 @@
-import * as algorithmData from '@/lib/services/data/algorithm.data'
-import { BreadcrumbShell } from '@/components/breadcrumb-shell'
-
-export default async function AlgorithmDetailBreadcrumb({ params }: { params: Promise<{ slug: string }> }) {
-    const { slug } = await params
-    const algorithm = await algorithmData.findAlgorithmBySlug(slug)
-
-    return <BreadcrumbShell items={[{ label: 'Algorithms', href: '/algorithms' }, { label: algorithm.name }]} />
-}
