+++ src/app/@breadcrumbs/models/page.tsx
@@ -1,5 +1,5 @@
 import { BreadcrumbShell } from '@/components/breadcrumb-shell'
 
 export default function AlgorithmsIndexBreadcrumb() {
-    return <BreadcrumbShell items={[{ label: 'Algorithms' }]} />
+    return <BreadcrumbShell items={[{ label: 'Models' }]} />
 }
