--- src/app/algorithms/[slug]/loading.tsx
@@ -1,9 +1,0 @@
-import { AlgorithmDetailSkeleton } from './_components/skeletons'
-
-/**
- * Loading shell for algorithm detail page.
- * Provides instant full-page layout before data loads.
- */
-export default function AlgorithmDetailLoading() {
-    return <AlgorithmDetailSkeleton />
-}
