+++ src/app/instances/[slug]/_components/server/model-detail-content.tsx
