+++ src/app/instances/[slug]/_components/server/model-detail-header.tsx
