+++ src/app/instances/[slug]/_components/server/prediction-run-table.tsx
