+++ src/app/instances/[slug]/loading.tsx
@@ -1,0 +1,9 @@
+import { ModelDetailSkeleton } from './_components/skeletons'
+
+/**
+ * Loading shell for model instance detail page.
+ * Provides instant full-page layout before data loads.
+ */
+export default function ModelDetailLoading() {
+    return <ModelDetailSkeleton />
+}
