+++ src/app/instances/[slug]/page.tsx
@@ -1,0 +1,43 @@
+import { Suspense, use } from 'react'
+import type { Metadata } from 'next'
+
+import * as modelData from '@/lib/services/data/model.data'
+
+import { ModelDetailContent } from './_components/server/model-detail-content'
+import { ModelDetailSkeleton } from './_components/skeletons'
+
+interface ModelDetailPageProps {
+    params: Promise<{ slug: string }>
+}
+
+export async function generateMetadata({ params }: ModelDetailPageProps): Promise<Metadata> {
+    const { slug } = await params
+    try {
+        const modelInstance = await modelData.findModelInstanceBySlug(slug)
+        const version = `v${modelInstance.versionMajor}.${modelInstance.versionMinor}.${modelInstance.versionPatch}`
+        return {
+            title: `${modelInstance.name} ${version}`,
+            description: modelInstance.description || `Details and prediction runs for ${modelInstance.name}.`,
+        }
+    } catch {
+        return {
+            title: 'Model Not Found',
+            description: 'The requested model instance could not be found.',
+        }
+    }
+}
+
+/**
+ * Model instance detail page showing model metadata and all prediction runs.
+ * Uses streaming with Suspense for instant layout render.
+ */
+export default function ModelDetailPage(props: ModelDetailPageProps) {
+    const params = use(props.params)
+    const { slug } = params
+
+    return (
+        <Suspense fallback={<ModelDetailSkeleton />}>
+            <ModelDetailContent slug={slug} />
+        </Suspense>
+    )
+}
