+++ src/app/models/[slug]/_components/client/copy-button.tsx
