+++ src/app/models/[slug]/_components/server/aggregate-performance.tsx
