+++ src/app/models/[slug]/_components/server/algorithm-detail-content.tsx
