+++ src/app/models/_components/skeletons.tsx
